webapi/webapi-serviceworkers-w3c-tests/serviceworkers/window-client-navigate.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>WindowClient.navigate()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="https://slightlyoff.github.io/ServiceWorker/spec/service_worker/index.html#client-navigate-method">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="w3c/service-workers/resources/test-helpers.js"></script>
<div id="log"></div>
<script>

promise_test(function(t) {
  var scope = "resources/blank-iframe.html"
  var scriptURL = "resources/client-navigate-blank-url-worker.js";
  var registration;
  return service_worker_unregister_and_register(t, scriptURL, scope)
    .then(function(registration) {
      return wait_for_state(t, registration.installing, "activated");
    })
    .then(function(state) {
      return with_iframe(scope);
    })
    .then(function(ifr) {
      frame = ifr;
      return new Promise(function(resolve) {
        var channel = new MessageChannel();
        channel.port1.onmessage = function(evt) {
          assert_equals(evt.data, "TypeError", "navigate url is 'about:blank'");
          resolve();
        };
        frame.contentWindow.navigator.serviceWorker.controller.postMessage(
          {port: channel.port2}, [channel.port2]);
      });
    })
    .then(function() {
      frame.remove();
      return service_worker_unregister_and_done(t, scope);
    })
    .catch(unreached_rejection(t));
}, "Check that navigate with 'about:blank' url will throw a TypeError exception.");

</script>
